requirements.txt
ultralytics
opencv-python
RPi.GPIO